rag/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
qdrant-client==1.9.1
rank-bm25==0.2.2
pdfminer.six==20221105
python-docx==1.1.0
reportlab==4.2.0
psycopg2-binary==2.9.9
pytest==8.2.0
